Added setup.py for easier installation, PyPI registration
Because I do not know the author or package maintainer, I had to make a
few assumptions about what the licensing, version, author, etc. were
from the existing code.

I'm fairly certain that I got it right, but it might be worth taking a
look. If everything looks good, this closes #8